+++ tests/Kiota.Builder.Tests/Extensions/OpenApiUrlTreeNodeExtensionsTests.cs
@@ -183,4 +183,51 @@ public class OpenApiUrlTreeNodeExtensionsTests
         Assert.Equal("{+baseurl}/{param%2Dwith%2Ddashes}/existing-segment{?%24select,api%2Dversion,api%7Etopic,api%2Eencoding}", node.Children.First().Value.GetUrlTemplate());
         // the query parameters will be decoded by a middleware at runtime before the request is executed
     }
+    [InlineData("\\reviews\\search.json", "reviews.search")]
+    [InlineData("\\members\\$ref", "members.ref")]
+    [Theory]
+    public void GetsNamespaceFromPath(string source, string expected)
+    {
+        Assert.Equal(expected, source.GetNamespaceFromPath(string.Empty));
+    }
+    [Fact]
+    public void GetsClassNameWithIndexerAndExtension() {
+        var doc = new OpenApiDocument
+        {
+            Paths = new(),
+        };
+        doc.Paths.Add("/reviews/{resource-type}.json", new()
+        {
+            Operations = new Dictionary<OperationType, OpenApiOperation> {
+                { OperationType.Get, new() {
+                        Parameters = new List<OpenApiParameter> {
+                            new() {
+                                Name = "resource-type",
+                                In = ParameterLocation.Path,
+                                Required = true,
+                                Schema = new() {
+                                    Type = "string"
+                                },
+                                Style = ParameterStyle.Simple,
+                            }
+                        },
+                        Responses = new OpenApiResponses() {
+                            {"200", new() {
+                                Content = new Dictionary<string, OpenApiMediaType>() {
+                                    {"application/json", new() {
+                                        Schema = new () {
+                                            Type = "string"
+                                        }
+                                    }}
+                                }
+                            }}
+                        }
+                    }
+                }
+            }
+        });
+        var node = OpenApiUrlTreeNode.Create(doc, Label);
+        var result = node.Children["reviews"].Children["{resource-type}.json"].GetClassName(new(){"application/json"});
+        Assert.Equal("ResourceType", result);
+    }
 }